added utils for handling join requests
